feat(bf):bf.madd
